+++ packages/app-component-manage/component-view-screen.tsx
@@ -13,6 +13,8 @@ import { ASSISTANT_PLUGIN_NAMESPACE__NOCHANGE } from "@core/constant";
 import { Edit, Settings } from "@material-ui/icons";
 import { ComponentCodebox } from "./component-codebox";
 import { ComponentSchemaEditor } from "./schema-editor";
+import { useSingleSelection } from "plugin-app";
+import { CodeBox } from "@ui/codebox";
 
 interface VisualComponentManifest {
   name: string;
@@ -22,8 +24,6 @@ interface VisualComponentManifest {
   gitUrl: string;
   codeSnippet: string;
 }
-
-import { useSingleSelection } from "plugin-app";
 
 export function ComponentViewScreen() {
   const [data, setData] = useState<VisualComponentManifest>(undefined);
@@ -65,7 +65,8 @@ export function ComponentViewScreen() {
 
   return (
     <div>
-      {/* <Preview auto /> */}
+      <Preview auto />
+      <CodeBox language="jsx" code="//aa" />
       {/* {selection ? (
         <form key={JSON.stringify(data)}>
           <p>component view placeholder</p>
