+++ packages/app-schema-editor/by-selection-state/selection-instance-component.tsx
@@ -1,0 +1,6 @@
+import React from "react";
+import { nodes } from "@design-sdk/core";
+
+export default function (props: { node: nodes.light.IReflectNodeReference }) {
+  return <></>;
+}
