+++ packages/app-schema-editor/by-selection-state/selection-invalid.tsx
@@ -1,0 +1,5 @@
+import React from "react";
+
+export default function () {
+  return <></>;
+}
