+++ packages/app-schema-editor/by-selection-state/selection-variant-instance.tsx
@@ -1,0 +1,5 @@
+import React from "react";
+
+export default function () {
+  return <></>;
+}
