+++ packages/app-schema-editor/by-selection-state/selection-variant-set.tsx
@@ -1,0 +1,10 @@
+import React from "react";
+import { nodes } from "@design-sdk/core";
+
+export default function (props: {
+  node?: nodes.light.IReflectNodeReference;
+  defaultComponent?: nodes.light.IReflectNodeReference;
+}) {
+  // TODO
+  return <p>select component inside variant set</p>;
+}
