+++ packages/app-schema-editor/index.ts
@@ -1,1 +1,2 @@
 export { SchemaEditor } from "./schema-editor";
+export { analyzeNode } from "./analyze-node";
