+++ packages/app-schema-editor/ux-messages/index.ts
@@ -1,0 +1,5 @@
+export const ERROR_MESSAGES = {
+  nothing_is_selected: "Nothing is selected",
+  you_must_select_instance_or_component_type_of_node:
+    "You must select instance or component type of node.",
+};
